backend/requirements.txt
Django==4.2.16
python-decouple==3.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
Pillow==10.4.0
django-filter==24.3
